% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGDP.R
\name{getStateGDP}
\alias{getStateGDP}
\title{Get a State and their associated Gross Domestic Product(GDP)}
\usage{
getStateGDP(state, save = FALSE)
}
\value{
GDP data.frame
}
\description{
Get a State and their associated Gross Domestic Product(GDP)
}
